+++ docs/CHANGELOG.md
@@ -5,7 +5,8 @@
     * Added `take_notification_hint(dict) -> bool`
 * **BREAKING CHANGE**: Changed DB schema of aggregated streams
     * Added columns; you will have to do a downtime deployment for these changes
-    * Use `python -m depeche_db generate-migration-script <PREV-VERSION> 0.11 --message-store=<NAME> --aggregated-stream=<STREAM-NAME> --aggregated-stream=<ANOTHER...>`
+    * Use the following command to generate a migration script:
+      `python -m depeche_db generate-migration-script <PREV-VERSION> 0.11 --message-store=<NAME> --aggregated-stream=<STREAM-NAME> --aggregated-stream=<ANOTHER...>`
 
 # 0.10.3